+++ internal/core/function_info.go
@@ -2,14 +2,16 @@ package core
 
 import (
 	"errors"
+	"fmt"
 	"github.com/matzefriedrich/parsley/pkg/types"
 	"reflect"
+	"strings"
 )
 
 type functionInfo struct {
 	funcType       reflect.Type
-	returnType     reflect.Type
-	parameterTypes []reflect.Type
+	returnType     types.ServiceType
+	parameterTypes []types.ServiceType
 }
 
 var _ types.FunctionInfo = &functionInfo{}
@@ -35,29 +37,39 @@ func (f functionInfo) Name() string {
 	return f.funcType.Name()
 }
 
-func (f functionInfo) ParameterTypes() []reflect.Type {
+func (f functionInfo) ParameterTypes() []types.ServiceType {
 	return f.parameterTypes
 }
 
-func (f functionInfo) ReturnType() reflect.Type {
+func (f functionInfo) ReturnType() types.ServiceType {
 	return f.returnType
 }
 
-func returnType(funcType reflect.Type) (reflect.Type, error) {
+func (f functionInfo) String() string {
+	parameterTypeNames := make([]string, len(f.parameterTypes))
+	for _, t := range f.parameterTypes {
+		parameterTypeNames = append(parameterTypeNames, t.Name())
+	}
+	funcTypeName := f.returnType.ReflectedType().Elem().String()
+	return fmt.Sprintf("f(%s) %s", strings.Join(parameterTypeNames, ","), funcTypeName)
+}
+
+func returnType(funcType reflect.Type) (types.ServiceType, error) {
 	numReturnValues := funcType.NumOut()
 	if numReturnValues != 1 {
 		return nil, errors.New("return type has to have exactly one return value")
 	}
 	serviceType := funcType.Out(0)
-	return serviceType, nil
+	return types.ServiceTypeFrom(serviceType), nil
 }
 
-func parameterTypes(funcType reflect.Type) []reflect.Type {
-	parameters := make([]reflect.Type, 0)
+func parameterTypes(funcType reflect.Type) []types.ServiceType {
+	parameters := make([]types.ServiceType, 0)
 	numParameters := funcType.NumIn()
 	for i := 0; i < numParameters; i++ {
 		parameterType := funcType.In(i)
-		parameters = append(parameters, parameterType)
+		serviceType := types.ServiceTypeFrom(parameterType)
+		parameters = append(parameters, serviceType)
 	}
 	return parameters
 }