+++ internal/tests/resolver_options_test.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"github.com/matzefriedrich/parsley/pkg/registration"
 	"github.com/matzefriedrich/parsley/pkg/resolving"
+	"github.com/matzefriedrich/parsley/pkg/types"
 	"github.com/stretchr/testify/assert"
 	"testing"
 )
@@ -20,7 +21,7 @@ func Test_Resolver_ResolveWithOptions_inject_unregistered_service_instance(t *te
 	r := resolving.NewResolver(sut)
 
 	parsleyContext := resolving.NewScopedContext(context.Background())
-	consumers, _ := r.ResolveWithOptions(parsleyContext, registration.ServiceType[FooConsumer](), resolving.WithInstance[Foo](NewFoo()))
+	consumers, _ := r.ResolveWithOptions(parsleyContext, types.MakeServiceType[FooConsumer](), resolving.WithInstance[Foo](NewFoo()))
 	assert.Equal(t, 1, len(consumers))
 
 	consumer1 := consumers[0]